+++ src/assets/Components/Novedades/Novedades.jsx
@@ -23,23 +23,11 @@ export default function Novedades() {
         setUsuarios(res.data)
     }
 
-    function parseJwt(token) {
 
-        const base64Url = token.split('.')[1];
-        const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-        const jsonPayload = decodeURIComponent(window.atob(base64).split('').map(function (c) {
-            return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-        }).join(''));
-
-
-        SetUsuariosession(JSON.parse(jsonPayload));
-
-    }
 
     useEffect(() => {
         getNovedades(),
-            getUsuarios(),
-            parseJwt(localStorage.getItem("token"))
+            getUsuarios()
     }, [])
